make more aggressive use of layer caching
this makes the images large for each component but dockers efficient filesystem
should mitigate this. And now all deps are cached in the base image.

Additionally, at the expense of not knowing the version inside the component
images, the .git repo (and so .git) is not longer part of the docker context and
so speeds things up